ENTERPRISE-728: Use base revision if no previous successful commit is available.
This was a problem when the build was first run for a new branch without any previous OK build.
In that case, the "GIT_PREVIOUS_SUCCESSFUL_COMMIT" env variable was set to null,
which lead to an exception in the Commit class constructor and early job termination.

The issue was introduced by recent refactoring (using Commit domain class instead of pure Strings).
However, even the former solution (with Strings) seems to be suboptimal,
because in that case following git command was used to get commit range:
```
git log --pretty='%%H' ..<current_commit_sha>"
```
, which lead to an empty commit range.